Dev mariam

+++ Controller/basicAdminController.js
@@ -10,7 +10,7 @@ const salt = bcrypt.genSaltSync(saltRounds);
 const basicAdminSchema = mongoose.model("basicAdmins");
 
 exports.getAllBasicAdmins=(request,response)=>{
-    basicAdminSchema.find({})
+    basicAdminSchema.find({}, {password: 0})
                 .then((data)=>{
                     response.status(200).json({data});        
                 })
@@ -37,6 +37,11 @@ exports.addBasicAdmin=(request,response,next)=>{
 }
 
 exports.updateBasicAdmin=(request,response,next)=>{
+    let password;
+    if(request.body.password){
+        password = bcrypt.hashSync(request.body.password, salt);
+    }
+
     basicAdminSchema.findOne({
         _id:request.body.id
     }).then((data)=>{
@@ -45,7 +50,7 @@ exports.updateBasicAdmin=(request,response,next)=>{
         }else{//for basicAdmin role
             
         }
-        if(request.file){
+        if(request.file && data.image){
             fs.unlinkSync(path.join(__dirname,"..","images",`${data.image}`));
 
         }   
@@ -54,7 +59,7 @@ exports.updateBasicAdmin=(request,response,next)=>{
         },{
             $set:{
                 lastName:request.body.lastName,
-                password: bcrypt.hashSync(request.body.password, salt),
+                password: password,
                 email:request.body.email,
                 birthdate:request.body.birthdate,
                 hireDate:request.body.hireDate,
@@ -71,9 +76,31 @@ exports.updateBasicAdmin=(request,response,next)=>{
 
 exports.deleteBasicAdmin=(request,response,next)=>{
     basicAdminSchema.deleteOne({
-        _id:request.body.id
+        _id:request.body.id,
+        isRoot: false
+
     }).then(data=>{
         response.status(200).json({data});
     })
     .catch(error=>next(error));
 }
+
+exports.numOfBorrowedBooks = async (request, response, next) => {
+	const curr = await bookOperattion.aggregate([
+		{
+			$match: {
+          operationOnBook:{$eq:"borrowing"},
+			},
+		}, // stage1
+      {
+        $group: {
+          _id: "$bookId",
+          Total_Number_Of_Operation: { $sum: 1 },
+        },
+      }, // stage2
+	])
+  .then((data) => {
+		response.status(200).json({ data });
+	})
+	.catch((error) => next(error));
+};
